Implement cloudinary

See merge request iseplife/api!30

+++ src/main/java/com/iseplife/api/entity/post/package-info.java
@@ -1,7 +1,7 @@
 @AnyMetaDef(
   name = "postEmbed",
   metaType = "string",
-  idType = "int",
+  idType = "long",
   metaValues = {
     @MetaValue(value = EmbedType.GALLERY, targetEntity = Gallery.class),
     @MetaValue(value = EmbedType.POLL, targetEntity = Poll.class),
@@ -13,9 +13,9 @@
 package com.iseplife.api.entity.post;
 
 import com.iseplife.api.constants.EmbedType;
-import com.iseplife.api.entity.media.Image;
-import com.iseplife.api.entity.media.Video;
-import com.iseplife.api.entity.post.embed.Document;
+import com.iseplife.api.entity.post.embed.media.Image;
+import com.iseplife.api.entity.post.embed.media.Video;
+import com.iseplife.api.entity.post.embed.media.Document;
 import com.iseplife.api.entity.post.embed.Gallery;
 import com.iseplife.api.entity.post.embed.poll.Poll;
 import org.hibernate.annotations.AnyMetaDef;
